requirements.txt
streamlit
feedparser
lxml_html_clean
readability-lxml
langchain
langchain_ollama